backend/requirements.txt
fastapi 
uvicorn 
pymongo 
python-dotenv
python-jose[cryptography]